+++ NEWS.md
@@ -3,7 +3,7 @@
 
 ## Major changes:
 
-- `osmdata_data_frame` adds columns `osm_center_lat` and `osm_center_lon` for `out * center;` queries (#316).
+- `osmdata_data_frame` adds columns `osm_center_lat` and `osm_center_lon` for `out * center;` queries (#316, #319).
 
 ## Minor changes