+++ NEWS.md
@@ -5,9 +5,8 @@
 
 - `osmdata_data_frame` adds columns `osm_center_lat` and `osm_center_lon` for `out * center;` queries (#316, #319).
 - Add parameters from `opq` to `opq_osm_id`: out, datetime, datetime2, adiff, timeout and memsize (#320)
+- Fix `available_tags()` function which no longer worked (#322 thanks to @boiled-data)
 - Implement `out:csv` queries (#321).
-
-
 
 ## Minor changes