+++ NEWS.md
@@ -1,4 +1,4 @@
-0.2.1.00x (dev version)
+0.2.2
 ===================
 
 ## Major changes:
@@ -12,7 +12,6 @@
 
 - Fix queries with `!match_case` and only one value (#317).
 - Fix queries with multiple features & multiple osm_types (#318).
-
 
 0.2.1
 ===================